fix: keep pk ids if supplied by users
