+++ lib/model.js
@@ -2383,7 +2383,9 @@ class Model {
         return this.QueryInterface.bulkInsert(this.getTableName(options), records, options, attributes).then(results => {
           if (Array.isArray(results)) {
             results.forEach((result, i) => {
-              instances[i] && instances[i].set(this.primaryKeyAttribute, result[this.primaryKeyField], {raw: true});
+              if (instances[i] && !instances[i].get(this.primaryKeyAttribute)) {
+                instances[i] && instances[i].set(this.primaryKeyAttribute, result[this.primaryKeyField], { raw: true });
+              }
             });
           }
           return results;
